try again to catch all the moved refs
