+++ catalogue_pipeline/id_minter/src/main/scala/uk/ac/wellcome/platform/idminter/steps/IdEmbedder.scala
@@ -5,8 +5,8 @@ import com.twitter.inject.Logging
 import io.circe.optics.JsonPath.root
 import io.circe.optics.JsonTraversalPath
 import io.circe.{Json, _}
-import uk.ac.wellcome.messaging.metrics.MetricsSender
 import uk.ac.wellcome.models.work.internal.SourceIdentifier
+import uk.ac.wellcome.monitoring.MetricsSender
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
 import uk.ac.wellcome.utils.JsonUtil._
