+++ catalogue_pipeline/id_minter/src/main/scala/uk/ac/wellcome/platform/idminter/steps/IdentifierGenerator.scala
@@ -2,8 +2,8 @@ package uk.ac.wellcome.platform.idminter.steps
 
 import com.google.inject.Inject
 import com.twitter.inject.{Logging, TwitterModuleFlags}
-import uk.ac.wellcome.messaging.metrics.MetricsSender
 import uk.ac.wellcome.models.work.internal.SourceIdentifier
+import uk.ac.wellcome.monitoring.MetricsSender
 import uk.ac.wellcome.platform.idminter.database.IdentifiersDao
 import uk.ac.wellcome.platform.idminter.models.Identifier
 import uk.ac.wellcome.platform.idminter.utils.Identifiable
